requirements.txt
flask
Flask-mongoalchemy
Flask-assets
flask-mongokit
Flask-Cache
Flask-OAuth
pytz